Match-id-749602d310aba7087a2f350500e5d5a228cc94a8

+++ src/core/initializer/random_normal_initializer/random_normal_initializer.h
@@ -19,7 +19,7 @@ namespace MxRec {
     class RandomNormalInitializer : public Initializer {
     public:
         RandomNormalInitializer() = default;
-        RandomNormalInitializer(int start, int len, float mean, float stddev, int seed, int initK);
+        RandomNormalInitializer(int start, int len, float mean, float stddev, int seed, float initK);
 
         ~RandomNormalInitializer() override {};
